+++ lib/core/widgets/shimmer_effect.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:shimmer/shimmer.dart';
+
+class ShimmerEffect extends StatelessWidget {
+  const ShimmerEffect(
+      {super.key,
+      required this.width,
+      required this.height,
+      this.borderRaduis});
+  final double width;
+  final double height;
+  final double? borderRaduis;
+
+  @override
+  Widget build(BuildContext context) {
+    return Shimmer.fromColors(
+      baseColor: Colors.grey.shade600,
+      highlightColor: Colors.grey.shade400,
+      child: Container(
+        width: width,
+        height: height,
+        padding: const EdgeInsets.all(8),
+        decoration: BoxDecoration(
+          color: Colors.grey.withOpacity(0.3),
+          borderRadius: BorderRadius.circular(borderRaduis ?? 14),
+        ),
+      ),
+    );
+  }
+}
