+++ lib/features/home/presentation/widgets/home_body.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:spotify_app/features/home/presentation/widgets/artist_card.dart';
 import 'package:spotify_app/features/home/presentation/widgets/home_tap_bar.dart';
 import 'package:spotify_app/features/home/presentation/widgets/news_songs.dart';
