+++ pubspec.yaml
@@ -53,6 +53,7 @@ dependencies:
   modal_progress_hud_nsn: ^0.5.1
   delightful_toast: ^1.1.0
   shared_preferences: ^2.2.3
+  shimmer: ^3.0.0
 
 dev_dependencies:
   flutter_test:
